Move `DatasetStreamProjection` from `mind` directory
