+++ src/core/algorithms/ind/mind/mind.cpp
@@ -11,10 +11,10 @@
 #include "algorithms/create_algorithm.h"
 #include "config/error/option.h"
 #include "config/names_and_descriptions.h"
-#include "dataset_stream_projection.h"
 #include "error/type.h"
 #include "ind/ind_algorithm.h"
 #include "max_arity/option.h"
+#include "model/table/dataset_stream_projection.h"
 #include "table/column_combination.h"
 #include "table/dataset_stream_fixed.h"
 #include "tabular_data/input_table_type.h"
