+++ src/core/model/table/dataset_stream_projection.h
